Fix parray index

+++ src/c/backend/include/parray_state.hpp
@@ -12,17 +12,17 @@ public:
   PArrayState();
 
   // Return True if there is an PArray copy (possibly invalid) on this device
-  bool exists_on_device(uint64_t device_id);
+  bool exists_on_device(int device_id);
 
   // Return True if there is an PArray copy and its coherence state is valid on
   // this device
-  bool valid_on_device(uint64_t device_id);
+  bool valid_on_device(int device_id);
 
   // set the exist status of PArray on a device
-  void set_exist_on_device(uint64_t device_id, bool exist);
+  void set_exist_on_device(int device_id, bool exist);
 
   // set the valid status of PArray on a device
-  void set_valid_on_device(uint64_t device_id, bool valid);
+  void set_valid_on_device(int device_id, bool valid);
 
   // get vector of valid devices
   std::vector<uint64_t> get_valid_devices();
